First version of serial code
The code has been imported from a Jupyter notebook and turned into a
Julia package. It should contain all the parts required for a DNS
channel flow, but the correctness of most parts has not been verified so
it should not be expected to give correct results yet. Much of the
functionality is not covered by unit tests at this point.