Parallelize solver with MPI
This is a major update that reorganizes much of how the code works. Each
module has been parallelized with MPI such that it is possible to run a
fully parallel DNS now. Also, most of the new code is covered by tests
now. As part of this rewrite, some implementatino approaches have been
changed to be better suited for the parallel version or to make the code
more clear and less error-prone in general.